requirements.txt
# Core
numpy>=1.24
opencv-python>=4.8
pillow>=10.0
openslide-python>=1.3.1
